Minor WIP Changes
There was a lot going on, and i want to make a checkpoint in my edits. Hard to organize ui commits in a format/template that works, between prefabs being split into directories that make sense, and unity inspector features and naming conventions.... there's a lot of back and forth.